requirements.txt

Django==2.1.3
djangorestframework
django-rest-auth
Pillow==2.2.1
dj-database-url
djangorestframework-jwt
django-cors-headers
#djangorestframework_simplejwt